Move ostree.remotes Angular module into src/app.js
With that, all Angular bits are in app.js. Change remotes.js to be a
standard ES6 module instead of an Angular module.

Closes #54
